replaced run with check output

+++ api/orchestrator.py
@@ -238,9 +238,11 @@ class Orchestrator:
     # Update linux and restart the device
     def update_restart_linux(self):
         Log.log("Updater: Update and restart requested")
-        subprocess.run(['apt','upgrade','-y'])
-        subprocess.run('reboot')
-        return 200
+        output = subprocess.check_output(['apt','upgrade','-y'])
+        if output:
+            subprocess.run('reboot')
+            return 200
+        return 400
 
     # Get all system information
     def get_system_settings(self):
